requirements.txt
PyPDF2>=3.0.0
Pillow>=9.0.0
pytesseract>=0.3.9
pytest>=7.0.0
